Day 2 part 1 complete, part 2 blech (#5)
my use of a list that I can pop on and off was not the best way to go
about this.

Other implementations looked good where if the first iteration failed to
check over two numbers, it would check all other combinations as well
before skipping the loop. wish i thought of that. instead i have a ton
of nested for loops.

Signed-off-by: Steven Billington <green2jello@gmail.com>